添加配置和运行脚本

+++ setup.sh
@@ -5,6 +5,7 @@ set -ex
 
 source ./env.sh
 
+
 # Install NGINX
 sudo apt-get -y install nginx
 # Disable the default website
@@ -12,26 +13,32 @@ sudo rm -f /etc/nginx/sites-enabled/default
 
 # set up the new website. To do that create the /etc/nginx/sites-available/onlyoffice-documentserver
 mkdir -p /etc/nginx/sites-available
-cp -f ${PWD}/conf/onlyoffice-documentserver.nginx /etc/nginx/sites-available/onlyoffice-documentserver
-sudo ln -s /etc/nginx/sites-available/onlyoffice-documentserver /etc/nginx/sites-enabled/onlyoffice-documentserver
+sudo cp -f ${PWD}/conf/onlyoffice-documentserver.nginx /etc/nginx/sites-available/onlyoffice-documentserver
+if [ ! -e /etc/nginx/sites-enabled/onlyoffice-documentserver ]; then
+  sudo ln -s /etc/nginx/sites-available/onlyoffice-documentserver /etc/nginx/sites-enabled/onlyoffice-documentserver
+fi
+
 
 # restart NGINX to apply the changes
 sudo nginx -s reload
 
 
-# Installing and configuring PostgreSQL
-sudo apt-get -y install postgresql
-sudo -i -u postgres psql -c "CREATE DATABASE onlyoffice;"
-sudo -i -u postgres psql -c "CREATE USER onlyoffice WITH password 'onlyoffice';"
-sudo -i -u postgres psql -c "GRANT ALL privileges ON DATABASE onlyoffice TO onlyoffice;"
-psql -hlocalhost -Uonlyoffice -d onlyoffice -f ${INSTALL_DIR}/documentserver/server/schema/postgresql/createdb.sql
+# Installing and configuring Mysql
+sudo apt-get -y install mysql-server
+sudo -i -u root mysql -uroot -p${MYSQL_PASSWORD} --execute="CREATE DATABASE IF NOT EXISTS onlyoffice;"
+sudo -i -u root mysql -uroot -p${MYSQL_PASSWORD} --execute="CREATE USER 'onlyoffice'@'%' IDENTIFIED BY 'onlyoffice';"
+sudo -i -u root mysql -uroot -p${MYSQL_PASSWORD} onlyoffice --execute="GRANT ALL privileges ON onlyoffice TO 'onlyoffice'@'%' IDENTIFIED BY 'onlyoffice';"
+sudo -i -u root mysql -uroot -p${MYSQL_PASSWORD} onlyoffice --execute="FLUSH PRIVILEGES;"
+sudo -i -u root mysql -uroot -p${MYSQL_PASSWORD} onlyoffice < ${INSTALL_DIR}/documentserver/server/schema/mysql/createdb.sql
 
 
 # Installing RabbitMQ
 sudo apt-get -y install rabbitmq-server
+
+
+# config fonts
 cd ${INSTALL_DIR}/documentserver/
 mkdir fonts
-
 LD_LIBRARY_PATH=${PWD}/server/FileConverter/bin server/tools/allfontsgen \
   --input="${PWD}/core-fonts" \
   --allfonts-web="${PWD}/sdkjs/common/AllFonts.js" \
@@ -48,3 +55,13 @@ LD_LIBRARY_PATH=${PWD}/server/FileConverter/bin server/tools/allthemesgen \
   --converter-dir="${PWD}/server/FileConverter/bin"\
   --src="${PWD}/sdkjs/slide/themes"\
   --output="${PWD}/sdkjs/common/Images"
+
+
+# Install Supervisor
+sudo apt-get install -y supervisor
+
+
+# Install bash scripts
+sudo apt-get -y install rsync
+rsync -avzrl --delete ./run/ ${RUN_DIR}/
+cp -f env.sh ${RUN_DIR}/
